fix Windows code for [bugs:#1840] lineout for very long lines is broken
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@12508 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ common/platform/windows/SysFile.hpp
@@ -129,7 +129,7 @@ public:
 
 protected:
     void   getStreamTypeInfo();
-    int    writeData(const char *data, size_t length);
+    int64_t writeData(const char *data, size_t length);
 
     int    fileHandle;      // separate file handle
     int    errInfo;         // last error info
@@ -156,5 +156,3 @@ protected:
 };
 
 #endif
-
-
